requirements.txt
langchain==0.1.0
sentence-transformers==2.5.1
chromadb==0.4.18
python-dotenv==1.0.0
streamlit==1.28.0
tiktoken==0.5.1
faiss-cpu==1.7.4
python-docx==1.0.1
PyPDF2==3.0.1
transformers==4.38.2
torch==2.2.1
accelerate==0.27.2 